refactor(listener): extract channel commands (#2074)
